clean up numpy deprecation, pyflakes warnings

+++ test/hl/test_committedtype.py
@@ -26,7 +26,7 @@ class TestCommittedType(TestCase):
         print("filename:", filename)
         f = h5py.File(filename, "w")
         # create a compound numpy type
-        dt = np.dtype([('real', np.float), ('img', np.float)])
+        dt = np.dtype([('real', float), ('img', float)])
         f['complex_type'] = dt
         ctype = f['complex_type']
         self.assertEqual(ctype.dtype.name, dt.name)
